Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -10,10 +10,10 @@ int main() {
         printf("%d",a-b);
     }
     else if (c=='*'){
-        printf("%f",a*b);
+        printf("%d",a*b);
     }
     else if (c=='/'){
-        printf("%f",(float)a/b);
+        printf("%d",a/b);
     }
     else{
         printf("error");
